PuzzleUpdate
The logic for the puzzle is there, at least for everything that can be
known at this point (missing stuff like proper scaling given screen size
and things of the like) however the need for a visual element is still
necessary. Saul, Ian, at this point I'm comfortable if you want to start
getting in there, either to add visual elements, or to add the logic for
adding tiles to the top of the puzzle when others get removed below it.