A start on visually dropping tokens
Split the method to drop tiles into 2 methods - one to do it visually
and one to do it in the beginning.
Added code that only allows the player to pick up tokens when in the
correct phase.
Mostly coded the visual drop - there are issues that need to be
resolved in ShiftTilesDownVisually().  Gameplay is unaffected.