Initial Enemy Spawner implementation
Began implementation for the enemy spawner. I am unable to test it at
current due to the bugs in the puzzle, which I will start to look into
now. Additionally, there will be more things I need to do in Enemy.cs in
order to get the arrays to be updated correctly when an enemy is
defeated, so I will work on getting that put in once Steven is done with
his work in that script.