Minor Refactoring
Added some code to make the re-position of the tokens after they fall a
little less obvious.  Instead of repositioning all of the tokens after
the shiftDown completes, they are repositioned the step after they
become too low.
Removed the loops that formerly did this, as they are no longer
necessary.