Added simple tutorial text
Tutorial text now displays when Show Tutorial is checked on the main
menu. Still needs a bit tweaking, but right now it shows one line of
text that disappears roughly after 8 seconds. Clicking on the text will
show another line. Also need to polish the look and placement
(currently shows through the title text). Will continue to work on
fixing the behavior so it actually guides the player along as they
perform actions.