Fixed health bar and Score
added different health bar asset (similar to the one in apocalyptia).
has text within it to indicate how much HP player has left. also fixed
the score so it doesnt cut off when it is 1000 or higher. changed
player health back to 100 since 200 felt a bit too much with the enemy
nerfs. plus it was easier to visualize the health when it was out of
100.