Health Tile + Intelligent Tile Tweaking
Added the Health Tile to the puzzle generator, which you can use and
combo with, also properly renders sprite. Furthermore, edited the
structure of the intelligent tile placement such that attack tiles drop
significantly more often, and the health tile was added to the mix.
Personal playtesting also leads me to believe the number of tile types
compared to the size of our grid is a little crowded, but as long as we
don't add any more we should be fine. Do not know if comoboing with the
tile actually heals you, as I cannot locate the health bar on the
in-game GUI.