tutorial animation test/sample
got the cursor sprite to move on top of the tokens for the tutorial. the
animations play right as the game starts and multiple can be placed on
the board at once. should be pretty simple to map them to a tutorial
board with differing animations. only thing thats missing is a looping
animation, that would be ideal.