Fixed the Move Queue centering
Made it so that the queue was being put off center by half the length
of the queue (as it is in Puzzle Manager), not by the whole length.  IE
added a “/2”.  =)
Also changed the moveQueue to a list, but that shouldn’t affect things.

Also added Temp/ to the gitignore.