Animations, timer, grayout, nerfs
Added in the player's new animations. Gray box appears over tiles when
player cannot make a move. Timer counts down to show how long the player
can hold a tile. Nerfed ranged enemy spawn. There is an odd error
message that pops up about Null Reference, and it seems to be related to
the animation changes, but it doesn't seem to break anything so I'll
look into fixing it later.