InversedPuzzleDroppingAlgorithm
In an attempt to correct the lack of tile variety issue, the opposite of
what occured before will now occur. Instead of giving you less of what
you use, you will get more. However if the player uses too much, they
will become single minded and only be able to do that one action,
effectively killing them off. Currently has no bounds but if we make the
player aware of it I can add them in soon. If we don't like it I made
changing it back a one-liner. We can talk about it tomorrow, but I want
to commit the function anyways.