More options, credits
added options to the options menu so you can toggle whether or not the
health displays numbers or not, and view the credits scene I put in the
other day.  Both the options and the credits are ugly but more work
depends on knowing what font we'll be using.  Will be fixed soon.