NewTilePlacementAlgorithm
Implemented Algorithm Cody suggested for Tile Placement, in current
build not the actual function being called, but the code is in there and
you just need to switch what placement function is being called, which
is easy enough to do, and after a little more bug testing tomorrow I'll
commit that as well. Good night everybody.