Death Fades and Animation
The player character, on defeat, now plays a death animation and then
triggers a fade to black before the high score screen.
NOTE:  The animation goes back to idle after the 3-tic animation;  I’m
gonna try to figure out a way to have it stay on the last frame of an
animation.