Updated volume sliders and options
Options text now scales.  I tested it on all the portrait screen sizes
that Unity offers, it looked terrible on the smallest, slightly wonky on
one of the weird ones, but great on all the common ones.  I also updated
the sliders for volume in the pause menu.  I don't intend to change them
any more before release.  Not as nice as they could look, but they are
functional and don't look terrible.